refactor: KitchenridersClient 를 도메인 패키지로 이동 - KitchenridersClient 를 사용하는것도 도메인 속성을 사용하는 로직이라서 도메인 패키지로 이동
