+++ src/main/java/kitchenpos/order/delivery_order/infra/DefaultKitchenridersClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.delivery_order.infra;
 
-import kitchenpos.order.delivery_order.application.KitchenridersClient;
+import kitchenpos.order.delivery_order.domain.KitchenridersClient;
 import org.springframework.stereotype.Component;
 
 import java.math.BigDecimal;
